+++ features/scripts/build_webgl.sh
@@ -16,16 +16,16 @@ popd
 pushd "$script_path/../fixtures"
 
 # Run unity and immediately exit afterwards, log all output
-DEFAULT_CLI_ARGS="-quit -batchmode -nographics -logFile build_macos.log"
+DEFAULT_CLI_ARGS="-quit -batchmode -nographics -logFile build_webgl.log"
 
 project_path=`pwd`/mazerunner
 
 # Build for Android
 
-$UNITY_PATH/Unity $DEFAULT_CLI_ARGS -projectPath $project_path -executeMethod Builder.MacOS
+$UNITY_PATH/Unity $DEFAULT_CLI_ARGS -projectPath $project_path -executeMethod Builder.WebGL
 RESULT=$?
 if [ $RESULT -ne 0 ]; then exit $RESULT; fi
 
-mv $project_path/mazerunner_macos.app $project_path/mazerunner_macos_$UNITY_PERFORMANCE_VERSION.app
+mv $project_path/mazerunner_webgl $project_path/mazerunner_webgl_$UNITY_PERFORMANCE_VERSION
 
-(cd $project_path && zip -q -r mazerunner_macos_$UNITY_PERFORMANCE_VERSION.zip mazerunner_macos_$UNITY_PERFORMANCE_VERSION.app)
+(cd $project_path && zip -q -r mazerunner_webgl_$UNITY_PERFORMANCE_VERSION.zip mazerunner_webgl_$UNITY_PERFORMANCE_VERSION)