+++ features/support/env.rb
@@ -74,7 +74,8 @@ BeforeAll do
     # Allow the necessary environment variables to be passed from Ubuntu (under WSL) to the Windows test fixture
     ENV['WSLENV'] = 'BUGSNAG_SCENARIO:BUGSNAG_APIKEY:MAZE_ENDPOINT'
   elsif Maze.config.browser != nil # WebGL
-    Maze.config.document_server_root = 'features/fixtures/maze_runner/build/WebGL/Mazerunner'
+    unity_version = ENV['UNITY_PERFORMANCE_VERSION']
+    Maze.config.document_server_root = "features/fixtures/mazerunner/mazerunner_webgl_#{unity_version}"
   elsif Maze.config.os&.downcase == 'switch'
     maze_ip = ENV['SWITCH_MAZE_IP']
     raise 'SWITCH_MAZE_IP must be set' unless maze_ip